feat!(ts): add typescript support
