+++ .prettierignore
@@ -1,0 +1,6 @@
+/node_modules
+/dist
+/.vscode
+/.github
+.gitignore
+yarn.lock
