+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+    "arrowParens": "avoid",
+    "printWidth": 80,
+    "semi": true,
+    "singleQuote": true,
+    "trailingComma": "all",
+    "tabWidth": 4
+}
