+++ src/constants.ts
@@ -1,0 +1,27 @@
+export default {
+    FORMAT_VERSION: 131,
+    NEW_FLOAT_EXT: 70,
+    BIT_BINARY_EXT: 77,
+    SMALL_INTEGER_EXT: 97,
+    INTEGER_EXT: 98,
+    FLOAT_EXT: 99,
+    ATOM_EXT: 100,
+    REFERENCE_EXT: 101,
+    PORT_EXT: 102,
+    PID_EXT: 103,
+    SMALL_TUPLE_EXT: 104,
+    LARGE_TUPLE_EXT: 105,
+    NIL_EXT: 106,
+    STRING_EXT: 107,
+    LIST_EXT: 108,
+    BINARY_EXT: 109,
+    SMALL_BIG_EXT: 110,
+    LARGE_BIG_EXT: 111,
+    NEW_FUN_EXT: 112,
+    EXPORT_EXT: 113,
+    NEW_REFERENCE_EXT: 114,
+    SMALL_ATOM_EXT: 115,
+    MAP_EXT: 116,
+    FUN_EXT: 117,
+    COMPRESSED: 80,
+};
