--- src/index.js
@@ -1,16 +1,0 @@
-'use strict';
-
-const Encoder = require('./encoder');
-const Decoder = require('./decoder');
-
-module.exports = {
-  pack: (v) => {
-    const encoder = new Encoder();
-    encoder.pack(v);
-    return encoder.buffer.slice(0, encoder.offset);
-  },
-  unpack: (v, { bigintToString = false } = {}) => {
-    const decoder = new Decoder(v, bigintToString);
-    return decoder.unpack();
-  },
-};
